PPI-Backend/requirements.txt
# FastAPI Framework
fastapi==0.117.1
uvicorn==0.37.0
gunicorn==21.2.0

# Database Core
sqlalchemy==2.0.43
alembic==1.16.5

# Database Drivers
aiomysql==0.2.0

# Environment & Configuration
python-dotenv==1.0.1
pydantic==2.10.5
pydantic-settings==2.7.1

# JWT Libraries
PyJWT==2.8.0
python-jose[cryptography]==3.3.0

bcrypt==3.2.0
passlib
email-validator==2.3.0

sendgrid==6.11.0
httpx==0.27.0
aiosmtplib==3.0.1
pytz==2025.2

# Redis client
redis[hiredis]==6.4.0

aioboto3==13.0.1
aiobotocore==2.13.0
python-multipart==0.0.6
Pillow==11.3.0
aiofiles==24.1.0

# Testing
pytest==8.1.1
pytest-asyncio==0.23.6
pytest-cov==5.0.0
pytest-mock==3.14.0
faker==24.4.0

pyexcel==0.7.0
pyexcel-xlsx==0.6.0
pyexcel-xlsx==0.6.0
openpyxl==3.0.10

qrcode==7.4.2
reportlab==4.2.5